Remove syncing state from Document Import
This removes 'importing' and 'syncing' states from
WhitehallMigration::DocumentImport as they don't provide much value,
particularly as we don't expect the import to be in these states for more than
a second or two. This also simplify our approach to retrying the import job,
which will be implemented in a future commit.