Retry Whitehall Document Import Job
We have decided to retry the Whitehall Document Import Job when a
`GdsApi::BaseError` is raised as we have a number of external GDS APIs which
could time out or fail for any number of reasons and they should be safe to
retry. At this point we've decided not to retry on a generic `StandardError`
exception as our current expectation is that this would most likely be raised
due to an issue in the document data or our code, in which case retrying would
be pointless and we'd be better off logging the error and retrying.  We may
decide to revisit this decision if we see a lot of these exceptions or they're
being raised due to issues that can be fixed by a retry.

We've decided to opt for 5 retries with the wait inbetween tries increasing
exponentially as we have some precendence for this approach in jobs such as
`EmailDeliveryJob` where retries are attempted after an API failure.