Validate document_types.yml against a document type schema
We want to iterate our document type tests to robustly check for
things like redundant keys, invalid values, etc. This is the
perfect task for a schema.
Most of the existing tests have been superseded by this approach,
apart from a check that the document type exists in 'GovukSchemas',
which isn't possible in a local schema validation.

I expect the schema itself to change quite a bit in the next
few weeks (for example, the addition of types to the enum array
in the 'content_field' property). We can now make those kinds of
changes more confidently and know that all of our formats
continue to work.

On @benthorner's suggestion I've split it into two 'sub-schemas':
"documentType" and "managedElsewhere", to be even more precise in
our validation for each type. The "managedElsewhere" one will be
removed soon, but this structure will be useful when validating
against different format types which may share similar fields.