Move "I'm not sure" feature test to request test
Avoid a slow and expensive feature test when the same thing can be
accomplished with a quick request test. I've deliberately used
a 'find' method on Supertype to retrieve the 'not-sure' ID, in
case it is ever removed in the future, otherwise the test could
pass while simultaneously not being a possible option anymore.

However, I didn't feel comfortable removing the feature test
without ensuring that we test that every Supertype is rendered
as an option on the "choose_supertype" view, as it's possible
we could filter out some options before rendering and therefore
not show a "I'm not sure" option at all. Our tests would continue
to pass, as if the user _were_ able to pass 'not-sure' to the
form, it would show the guidance as expected, but in reality the
feature would be broken.