Use symbol keys when generating API payload
This makes the way we generate the overall payload for the Publishing
API consistent with the syntax we use for the parts of it that are now
generated by inidividual Field classes. To avoid knock-on changes, I've
retained the string keys for the payload output, which is in any case
a serialisation concern, and shouldn't prevent us using shorter syntax.