Remove 'revision' and 'content' param groupings
Previously we grouped params on the content screen in order to make
extraction convenient, which is no longer relevant since most of them
are now extracted by individual fields. The use of 'require' is now
irrelevant, since there's no danger of 500s from trying to access a
nested hash (revision: {}, contents: {}) that doesn't exist.

Note that using the more conventional 'params.permit' for change note
params would generate a warning in the presence of other params like
'title', since the expectation is that all (valid) params are permitted.