Remove and_the_preview_creation_succeeded steps
Previously we added these steps to many features as a way to check
the tangential outcomes of a user action, including the document
state and API requests to Publishing API and Asset Manager, which
arise from the PreviewDraftEditionService. Checking API requests
in features was pragmatic when we had limited tests at lower levels,
and the idea was that each feature would cover its specific part
of the Publishing API payload. Since then, we've introduced the
'failsafe preview' feature, such that its far less important or
'optional' that a request is made, since a fallback exists. The
checker for draft user state was added with a similar mindset, to
ensure that the edition state is (re)set appropriately on edit,
which is now covered by 'editable?' assertions.

This removes the checks for API requests and draft user state, and
renames or removes the associated step method to leave one that
checks for a timeline entry only.