Ignore change notes for first editions
This adds a check to ensure we don't allow updating a first edition
with a custom change note (or update type). Since the impact of this
happening is relatively minor, we avoid raising an error and silently
drop the extra parameters if they were somehow passed. Since we can
now expect the presence of the change_note and update_type params for
non-first editions, it's no longer necessary to 'compact' them away.

Changing the test setup to support the extra 'change_note' param was
made easier by replacing the 'build_params' method with the use of the
inbuilt 'merge!' method on the ActionController::Parameters class.