Stripped 'news_article_spec.rb' down to concentrate on previewing
We only care that a user can create a document from scratch and take
the minimal steps necessary to make the document valid against its
schema. This validation alone should give us confidence that it can
preview (and publish) correctly. Other functionality, such as tagging,
is handled in its own feature test and does not need to be re-tested
here.

Changes:

- Checks are made against Document.last, rather than Document.first,
  as this makes more sense
- Avoid doing any tagging, as this isn't required to preview the
  document. Also avoid stubbing role appointment links, as these
  aren't required for a minimum viable document
- Remove 'reset_executed_requests!' call, as only one call is now made
  and therefore we don't need to disambiguate
- Remove publishing API stubbing in the 'before' block, as no tagging
  is done now.
- Removed 'summary', as this isn't strictly necessary for a
  previewable document