Strip exif data from imported images
We need to strip exif data from images we import due to whitehall not
acknowledging these attributes, for example - rotation.

This causes problems in the browser if an image has rotation exif data
as in certain circumstances the browser will display the image rotated
(image open in new tab) and sometimes not.

The problem also persists when we import a rotated image to content
publisher as we respect the orientation before we strip the data which
ends up potentially rotating the image in a way the user didn't expect
which then fails validation, E.g width = height.

To get round this issue we want to strip the image first to ensure
content-publisher sees the image in the same way whitehall does.

Trello:
https://trello.com/c/b3vqfsA5/1395-deal-with-rotated-images-in-import