Update import single Whitehall document task
Now we have got rid of `WhitehallImporter#import_and_sync` we need to remove it
from the rake task to import a single Whitehall document. For now we replace it
with calls to `WhitehallImporter::Import` and `WhitehallImporter::Sync` and
we've copied rescuing any errors raised by these wholesale from the
`WhitehallDocumentImportJob`. This is a temporary measure; we intend to modify
this rake task to make use of `WhitehallDocumentImportJob`, however some
additional work needs to be done to achieve this, which will be done in a
subsequent PR.