Remove check_path_conflict_flag for base paths
Previously we added this flag to avoid having extra stubs in many
tests, due to the API call we make in order to check a base path is
(probably) available. This removes the flag, so that the check is
always enabled, noting that we should be able to use the new flexibility
of the TitleAndBasePathField to achieve the same effect.