Remove unhappy edge case steps in feature
Previously we checked that editing an edition under review did not
revert its state to 'draft', which isn't a core aspect of the 'Submit
for 2i' feature. Our code is a long way away from allowing this to
happen, noting that we only reset the 'draft' state in one place.
This removes the extra steps, as they're not clearly adding value.