DRY-up specs needing a 'body' field
This adds a 'with_body' trait to DRY-up all but 4 places in the tests
where we need to explicitly set 'contents' fields on a document type. I
did consider adding a convenience syntax instead:

   content_field_ids: %w(body)

Since this would be longer to type, and potentially introduce confusion
about how the factory works, I thought it would be simpler to just use a
trait to cater for the majority use-case.