Use more keyword args for services
This introduces and renames some of the arguments across a few services,
in order to increase readability at the point of use. We agree that it's
not necessary to use keyword args everywhere, as they only seem to add
value where the arguments passed are not self-explanatory:

   nil, false, true, <some_expression>

Some of the arguments renamed include:

   - GenerateBasePathService: proposed_title -> title
   - GenerateUniqueFilenameService: suggested_name -> filename
   - PublishAssetsService: live_edition -> superseded_edition

In the latter case, the previous name of the 'live_edition' argument
conflicted with the resync use case for this service, where the live
edition is passed as the first arg (and the second left blank).