Remove 'revision' param for edition checkers
Previously we passed the 'edition' and the 'revision' between checkers
for requirements issues, often defaulting the 'revision' to the current
one for the edition. Having both objects available lead to confusion
about which to use to access data, and was originally because the
revision had updated data and was not (yet) assigned to the edition.