Display imported fact checks from Whitehall
Internal notes with content are highlighted to the user, those without content
are not.

There are two things we need to maintain with respect to fact checks: a request
(i.e. a request sent to a department or individual) and a response (i.e. the
response coming back from the recipient(s) of the original request). These are
migrated as two seperate entries.