Update edition before we check for issues
This ensures an edition has updated data by the time it gets passed
to requirements checkers, which resolves the original issue that lead
to passing the revision as a second parameter. While assigning the
new revision is technically unnecessary, taking this extra step before
checking for issues seems to be a good compromise and shouldn't impact
other aspects of processing the request.