Update use of user_ids and centralise lookup for consistency
The 'user_ids' hash is constructed by mapping a Whitehall user id to
the same user's id in Content Publisher via their uid.  For example:
Fred Flintstone has a uid of "dce8e3a3-2872-4d45-848e-d1012d8cfd89"
in both Whitehall and Content Publisher.  However, their id in
Whitehall is 615 and in Content Publisher it is 27.  The resulting
hash to map this user would then be '{ 615 => 27 }'. Note that the
key is the Whitehall user id.

On migration the various time line entries from Whitehall need to be
mapped to their Content Publisher equivalent in order to avoid
miss-matched or nil user id's on migration.

This change ensures that the mapping between Whitehall and Content
Publisher user is done in one place: inside the 'create_timeline_entry'
method, and that all user id's passed into that method are
Whitehall user ids.

The appropriate parameter to this method has also been renamed
for clarity.