Unlock a Whitehall document if import fails
If an exported Whitehall document fails to be imported, we need to unlock it so
that users can continue to edit it in Whitehall. This shouldn't happen if the
document has been imported but hasn't been synced.  We also move some tests
that depend on whether the document has failed to import or sync to sit under
the new contexts we've created, as they don't cover behaviour that is specific
to a StandardError being raised.