Use faster Capybara matchers to check absence
Previously we used the Capybara 'all' selector to count the number of
assets on a page and see it was zero. Using 'all' takes a while because
it tends to wait for the entirety of Capybara.default_max_wait_time,
since Capybara wants to be sure the page has finished loading. This
changes the selector to explicitly look for what we're trying to check,
which only needs to wait until the condition is true.

Note that we need to put a positive assertion before the negative check,
to make sure we've actually switched to the page we're trying to check.

https://github.com/teamcapybara/capybara#asynchronous-javascript-ajax-and-friends