Only use JS where it's essential to the feature
Previously we enabled JS for a bunch of features where the test for
the no-JS fallabck is actually the same; the question is only whether
the associated responses will also work in a modal context. This
changes our features to use the faster, default 'Rack Test' driver,
unless the JS driver (Chrome) is essential.

We now have Jasmine tests to cover each modal workflow, and a single
place where the appropriate layout is set, which ensures any page will
render correctly within a modal. This change does reduce coverage in
one respect: the action overrides on links and buttons, which we could
think about having a view test for.