Set layout globally from rendering_context
Previously we set the layout explicitly for actions that can render in a
'modal' or 'application' context. It's hard to test the layout is set
correctly, since we would need to look for specific elements in the
response body. Instead, this DRYs-up by making the layout for all pages
sensitive to the rendering context, reducing the hole in tests.