Add tests for Editions::DestroyInteractor
The impetus to add tests for this was learning that there is actually a
problem where the edition model used in this interactor is not up to
date and thus a timeline entry is created with the wrong status.

This adds tests for the whole file however the one on line 20 is a
failing test that captures this problem.