Use edition as subject of DeleteDraftEditionService
This changes this service to take an edition as it's argument rather
than a document. This change has been done in part to resolve a problem
where operating on a document meant a stale edition was used in the
interactor. When trying to address that problem I noticed that it seemed
quite strange that this class operated on a document when its name
implies an edition.

I also added reload methods for associations of a document so that were
a similar problem to occur again the document is up to date.