Rearrange and refactor DeleteDraftEditionService tests
I found these tests quite hard to follow as there are quite a lot of
them. I decided that it would be easier to work with them if they were
in certain groupings so I've moved them into individual describe blocks.
This might well be a sign that this class has too many responsibilities
but that seems like a problem for another day.

Aside from these I've adapted these tests to be much more edition
centric and to no longer rely on any reloading. I've had to tweak a
bunch of factory calls and to resolve some inconsistencies with use of
brackets or not I've made them all use brackets - hopefully this isn't
too controversial.