Rename alternative_path to alternative_url
It turns out that the Publishing API naming of alternative_path is
misleading as this value can be a full URL and doesn't necessarily need
to be a path. Since a path is a type of URL it seems better to use url
in the name then both types of value are semantically correct. This is
also the naming used in Whitehall so has consistency with that.