Adjust remove rake task tests
This makes the tests more focused on what the tasks do rather than
consequences that happen at deeper layers (API calls). In order to
maintain the underlying code being actually called I've put in a
`and_call_original` so there is an execution of the method even if we
don't assert on things the method does.