Fix link handling in removal timeline entry
This fixes a bug where we had nil set on the wrong argument in link_to -
it needs to be the first argument if we want the link text to match the
href. It also resolves a problem where we only expected the
alternative_url to be a path so suffixed it on the hostname, to resolve
this we use URI.join which will default a scheme and hostname from Plek
and overwrite those based on the value of alternative_url.

As this was a bit fiddly and something that was broken I've added tests
for it.