Refactor checkers to distinguish the data to check
Previously we initialized some of the checkers with the data to check,
while in other checkers the data is passed to the specific method where
we check for issues. Sometimes this is reasonable, if the checker could
only have one use (e.g. FileUploadChecker), but most checkers could have
multiple ways of checking issues. This refactors the checkers so that we
consistently pass the data to check to the individual issue methods.