Replay form submission with params
Previously we relied on the in-memory revision to store and replay
form parameters if we need to show requirements issues on the edit
page for contents and tags. This changes the pages to use a more
conventional approach: params[:field] || value_from_db.

Making this change will support the following work, where we stop
trying to re-use a revision as a temporary store for parameters,
which has lead to some confusion and inconsistency in interactors
and requirements checkers.