Use params directly in requirements checkers
Previously we used the in-memory revision to check for requirements
issues, which is confusing because it wasn't clear why the revision
should have the updated/dirty data. This refactors the checkers for
contents and tags to use params directly, and otherwise rely on data
from the stored edition e.g. for pre-publish issues.