Use I18n for phrasing of remove entries
This text was inconsistent with other text in this view as it wasn't
sourced from the I18n file. It's not the most friendly of changes as
the phrases are so short that the I18n lookups use a lot more space.

I couldn't think of names for the keys that didn't essentially match the
text since the text is so contextual.