Check for issues before updating the revision
Previously we had to update the revision in order to check it for
issues, which was somewhat confusing, as intuitively we only want
to perform an 'update_revision' if there are no issues. This moves
the check for issues before the update for revision

Previously the interactor for tags spread the use of its updater
across two methods. This refactors the interactor to follow the
approach we take elsewhere for updaters, which avoids making it
accessible between methods.