Revert "Avoid 'feature' and 'scenario' in specs"
This reverts commit 9e99cf3758849a4529f213684d0c2315308aa284.

After some discussion, we agreed that these keywords make the most
sense in a feature context, in place of arbitrarily-worded 'it'
blocks. One option we considered was making this rule non-directory
specific, which would mean whitelisting the 'feature' and 'scenario'
keywords; this seemed like a worse choice, as long as we're content
that the 'spec/features' directory is a common convention.