Send asset ID to Asset Manager
Previously we were getting an ID from the `original_asset_url` exported with an
image from Whitehall and sending this to Asset Manager in order to make changes
to the asset. This isn't the right ID however, and resulted in Asset Manager
being unable to find the asset. What we actually need is to find out the Asset
Manager ID and send that to Asset Manager. We can do this by getting the legacy
URL path from the `original_asset_url` and looking this up in Asset Manager to
get the correct Asset Manager ID.