Fix image integrity bugs
Strings were previously being used to compare the attribute when the
attribute was always a symbol, the strings have now been converted to
symbols.

There was multiple reasons these issues were not caught by the tests:

For the `default_image?` scenario the reason was because the 'image
problems' logic would never be called due to the if statement -
`publishing_api_image[attribute] != proposed_image_payload[attribute]`
reporting false due to a missed level of nesting in the
publishing_api_content stub, `[:image][:caption]` instead of what it
should be `[:details][:image][:caption]. This caused a false positive to
be returned.

For the `empty_caption?` scenario the reason was due to the
document_type not being set to accept images and the edition not having
a lead_image_revision set, this caused the
PreviewDraftEditionService::Payload to not acknowledge the edition had
any images, this combined with the incorrect stub caused the 'image
problems' logic to not be called again and record a false positive.