Dry up repeated let blocks
We repeat the same grouping of lets in every context of this test. This
change pulls them out so they can be shared at a higher scope. This does
impact the distance from usage of the let, however I believe this to be
an overall net benefit to the readability of the code given how similar
the contexts can look with such repetition.