Move schemas to spec/support/schemas
Schemas aren't used at runtime, so it makes sense to move them
under the spec/ directory.

I did consider moving it to `spec/support/apis/schemas`, which
would allow us to remove a line of config in spec_helper (since
this is the default location for schemas), but as we're not testing
an API endpoint it makes little sense to keep it there.