Add change_history to metadata_revisons
We want to create a column for a revision that can store the change history
for that edition. By change history we mean an array of all the past change
notes with their times.

Created a new column on 'metadata_revisions' called 'change_history'.  Column
type is JSON, default is [] (empty array), and is not nullable.