Add validation for change notes stored in MetadataRevision
'change_history' is a JSON column that has been added to MetadataRevision
and defaults to an empty array.

Change notes will be added to this column.

Each change note should be a JSON object with 3 attributes:

* 'id' - A uuid that we generate
* 'note' - the change note content or text
* 'public_timestamp' - a string representing the RFC3339 representation
of the 'edition.published_at' timestamp

To have confidence in this JSON structure, we should add a validator on
MetadataRevision that tests both the structure and that the fields are
what we expect them to be (e.g. 'id' is a valid UUID and
'public_timestamp' is a valid RFC3339 representation).

The validator is defined as 'strict'.  This ensures a particular
exception is raised and that the application will crash with suitable
debug information. 'strict' validators differentiate from the type of
validation that is used to give user feedback.