Replace generic TagField class with specific ones
This moves us towards a 'flexible fields' approach, whereby tag
configuration is more declarative and can be kept consistent
between different document types. It avoids the possibility of, say,
configuring a 'primary_publishing_organisation' to be 'multi_tag'.

It involved changing the schema, obviously - I took the opportunity
to tidy up the schema where it makes references to definitions
that aren't reused, and could be defined in place. There's also
little benefit in using enumerated values for the tag/content
fields since the tests blow up if unrecognised fields are added
anyway.

It also involved removing the factory, as it is not needed - we
can simply instantiate the fields directly. The alternative would
have been to pass traits to specify which tag field class to
instantiate - something that was difficult to enforce - and we
weren't using any of the other aspects of factories to warrant
their use.