Make contacts code more resiliant
Now when there are no contacts for an edition the embed is ignored, the
same is true for when there is no matching contacts for an embed.

Previously the code would throw an error at either of these two
scenarios listed above as it tried to look for a content-id on nil.

Trello:
https://trello.com/c/OzdqkkW2/1455-fix-documents-that-failed-the-import-with-an-array-error