Define pre_update_issues on a per-tag-field basis
Following the same approach as content fields, we have consolidated
the business logic into each respective tag field type rather than
higher up in the chain.

Despite not needing the 'edition' parameter, I have passed this
(as well as the 'params' parameter) to maintain consistency with
the content fields, so that we can further consolidate them in
future.