Make "path" required if "type" is "managed_elsewhere"
If the format is not one that can be created in Content Publisher, we need to route users to
the correct publishing app or piece of guidance. To be able to do that we need to make
"path" a required field, but only if the type is "managed elsewhere".

There is is no elegant way to do this.
I investigated using the new `if-then-else`[1] logic, or trying to set `dependencies[2],
but neither of these approaches failed schema validation when a "path" was removed from
`config/document_type_selections.yml`, so I went with the only approach that did work:
splitting up the "type" enum and creating two schemas[3] to match against.

[1]https://json-schema.org/understanding-json-schema/reference/conditionals.html
[2]https://json-schema.org/understanding-json-schema/reference/object.html#dependencies
[3]https://tonyyan.wordpress.com/2015/03/12/json-schema-based-on-property-values-use-of-oneof/