Backfill change_history for metadata_revisions
This migration backfills the change_history for all metadata_revisions,
the change_history is built up from the individual change notes from
each edition where they have been published with a major change.

The change_history for an edition is built up from the previous editions
change notes and consolidated into the change_history field on the
corresponding metadata_revision. Not all change notes should be included
in this history however, the first published change note and the current
change note should be excluded from this list. E.g for the 4th edition
of a document it would contain the change notes from the 2nd and 3rd
edition but not the 1st (first published) and 4th (current).

Each change note included in the change history will include a UUID,
public_timestamp (publishing time) and the change note itself.

Trello:
https://trello.com/c/ymfW8Eeh/1462-backfill-the-changehistory-of-existing-content