Remove enum matches from document_type schema
These enums don't add any benefit. They just add to the burden of adding
new document types.
The "id" doesn't really matter, because it's not user facing, but has
been changed to only allow letters and underscores to match the document
type selections schema.
The "schema" is already being checked against the "source of truth",
`GovukSchemas::DocumentTypes.valid_document_types`, in the tests, so we
don't need to also maintain that list here.