Rename 'images' attribute to 'lead_image'
Previously we toggled the lead image feature on document types with an
attribute called 'images', which was originally accurate before inline
images became a feature. This renames it to better reflect its function.

Rather than renaming the attribute in the tests, I chose to use a trait
to decouple the toggle mechanism from the use of a custom document type
across the image specs, noting the repetition we used to have.

One of the feature specs was redundantly specifying the 'images' flag
when testing inline images. This moves the toggle for the lead image
feature into the correct scenario (editing lead images).