Add new assertion for edition features
This introduces a new 'assert_edition_feature' method and applies it
to lead image actions in order to render/return a 404 response when
the user tries to perform a lead image action on an edition that does
not support that feature, as specified in its document type.

I chose to make the new assertion similar to the one for edition state,
with the compromise that the usage is like '&:lead_image', instead of
just ':lead_image'. The problem with using a symbol alone is that it's
unclear which symbols are permitted, and what to do if an invalid input
is passed. Using an extra '&' character to simplify the implementation
and gives us flexibility to do more complex feature assertions.