Move deleting a lead image into scenario
Previously the feature test for deleting a lead image was separate
from the other deletion scenarios. While the outcome is slightly
different, the user actions are so similar that it sets a confusing
precedent to have it in a different file. This incorporates the
scenario into the general feature test for deleting an image.