Create service object to encapsulate the logic for creating a new edition
When creating a new edition we first create a new revision.  This new
revision needs to hold the complete change history of the document -
including the change note(s) of the edition that is being superceeded.

When deciding to add the current edition's change note(s) to the overall
change history, we use the following criteria:

* The current edition's update type is major
* The current edition has a change note
* The current edition is not the initial first published one

Once the new revision has been created, we create the new edition itself
and link the revision to that edition.  In this way the document change
history gets updated and copied into each new edition.