Introduce style constraints in rubocop
It was raised in a previous pr (https://github.com/alphagov/content-publisher/pull/1748#discussion_r377259573) that we should have a linter to look at TimeZone. In that discussion it was stated that we should prefer `Time.current` over `Time.zone.now` but this is not a thing that rubocop supports. Instead I've implemented the opposite in order to ensure that we have a consistent style.