Manage order index
In a no-JavaScript experience, users will see a set of index numbers provided by the backend. They will be able to edit these values and submit the form for a reorder. When enhanced with JavaScript, these inputs will be automatically updated after each drag/drop or up/down click to reflect the current order.