Use consistent naming for button to crop image
Previously we used custom 'Crop image' text when passing through the
'crop image' page. This updates the text to make it consistent with
the upcoming work on file attachments (as we add an extra edit page).

When no crop is occurring, we still say 'Continue'. The 'Crop image'
text may have been intended to reassure the user of the outcome of the
crop, but we think that 'Save' should accomplish this just fine.