DRY-up `pre_release` configuration
Having to define `pre_release` in two places is a risk as it means they
could become out of sync.

`pre_release` information is really about the state of the document
type, and not navigation, so document type selection now gets this
information from  DocumentType.

NB: I've tried to stub these tests as much as possible, so the tests don't need
to be updated every time the config changes, however because the
content on the page is pulled from translation files, I've had to
set the `id` of the stubbed document type to "news" because it
already has an entry in the document_type_selections locale file.