Check if user can access pre_release document
The ability to create new publications, or filter by publications on the
index page has been restricted to users who have the
"pre_release_features" permission.

We need to check that only users with these pre_release permissions can
view or edit publications documents as well.

This is so that we don't end up with users being in the position of
being able to do some parts of pre-release format work based off a link
but not being able to create one or find it in the index.