Move check for pending document import before importing
We originally checked if a document import is in a pending state within the
WhitehallImporter::Import class. However we've found that it's beneficial to do
this check before the class is called, as in situations where a document has
successfully been imported but the job is being retried as a result of an error
in the syncing process, the job errors since the document import is no longer
pending.