Exclude pre-release publications from edition filter results
Trello - https://trello.com/c/wzzVIihM/1521-exclude-prerelease-publications-from-edition-filter-results

As it stands, users without the pre_release_features permission are able to see editions with the pre release flag when they visit content-publisher. This will give the false impression that these are things that can be edited. This adds the functionality to remove these editions from the list of things that such users can see.