Keep track of change_history_ids during backfill
This applies a fix to the previous version of this migration. Previously
when you ran this migration the ids of each change history entry would
change for every MetadataRevision. This meant the revision history
showed lots of tweaks to change_history.

To resolve this we keep track of a change history id that corresponds to
an edition id. Anytime we work with the same edition we can re-use that
id. When running this migration we only see entries in the revision history
for when a new item is added to the change history.