Save the featured attachment
This reuses the existing functionality to create an inline attachment.
So that the "create" controller knows which page to display if there are
any issues, a "wizard" param has been added to the "new" page.

If the controller sees this "wizard" param, it knows that the user is in
the featured file attachment flow, rather than the inline attachment
flow.