Expose new 'attachments' attribute
This adds a new 'attachments' method to document types, which exposes
the value of the attribute with added question methods e.g.

    document_type.attachments.featured_and_inline?

This should make it shorter/simpler to check the value of the attribute
as we add conditionals that depend on it, and is also consistent with
the existing boolean question methods e.g. for 'pre_release?'.