Fix incorrect ordering of change history migration
Oh dear - I only noticed this when this migration had been run in
production. It's not too big a concern as we don't validate that the
change history is in a particular order and can likely recover from it
being out of order.

However since this is a backfill it'd be good to set the precedent to be
in the order we expect.

To resolve this I plan to rollback the migration in each environment
(which won't do anything) and then re-run it to reset the change
history. It doesn't seem to be a sufficiently big problem to warrant
creating a new migration.