Make tag and content fields take the same view parameters
This enabled me to remove the 'document_type' method from the tag fields.
I had to change the 'rescue GdsApi::BaseError' to a 'ActionView::Template::Error'
as the error is now raised from within a sub-partial, which causes a wrapped
error.
I also remove the `tags[*]` outer portion of the `name` attribute,
as we have done for content fields.