Use consistent describe blocks in request specs
Previously the 'errors' spec was inconsistent with all other request
specs, and with itself: half of the file used a consistent format

   describe "METHOD /path-under-test"

while the other half used freetext. I found this confusing when
reviewing a change to this file, as the request(s) under test was
unclear. This modifies the 'errors' spec to use a consistent format
around each test, using just enough freetext to indicate the range
of paths being described (even if they're not actually covered by
the test).