Backfill removed_at value for Removals
In anticipation of making `removed_at` a required attribute for a Removal
record, we backfill the value for existing Removals. Since we set the
`removed_at` value to the current time when a Removal is created, it makes
sense to set `removed_at` for existing Removals to their `created_at` value.