Implement importing change history
This commit makes use of the ChangeHistory class to generate change
history which can be saved on the editions metadata revision. We've had
to handle the creating of revisions differently to accommodate the
change as it's now possible that there can be two distinct
metadata_revisions per whitehall edition (if it has been unpublished and
edited).