Keep existing removed_at when removing edition
In cases where we attempt to remove an edition that has already been removed
(for example, if the explanation needs to be updated) then we need to ensure
that the existing `removed_at` date is kept the same. This modifies the Remove
Document Service to set `removed_at` and handle whether it should be the
current time or the existing value. It also means we can remove setting this
from the remove rake tasks.