Re-order db/schema.rb
This document_subtypes column is added through a migration, so ends up
at the end of the column list.  However, due to a un-noticed merge conflict
in aa31f36e2s, it ended up in the middle of schemas.  This causes issues
each time a developer runs a migration, since Git detects a change in the
db/schema.rb file.