Some subtle refactoring of MetadataRevision validation specs
We were unnecessarily hitting the db for each of these so I switched to
build. I also didn't think the failure_prefix variable offered us much
in the way of value since it was similar in length to what it
represented. The describe seemed to be named slightly wrong too.