Handle API errors within each tag field view
This allows us to revert back to the more specific 'GdsApi::BaseError'
exception handler, and makes each tag field view more transferable
(robust and can work in isolation).

A side effect of this is that the fallback message will be shown
several times if the API is down: one for each tag field. In
theory there could be an API failure when retrieving some of the
fields, but a valid response when retrieving the others, which
would make this an improvement as we're at least showing the data
we've been able to retrieve. In practice, if the API is down it is
likely to be down for all fields. We may want to consider shortening
the error message.
See full discussion here:
https://github.com/alphagov/content-publisher/pull/1804#discussion_r387047128