Wrap Resync Service in a Sidekiq job
When the resync service is run on one document, it generally completes
without any errors. However, when rake task `resync:all` is run, and
every document in Content Publisher is resynced, the task more often
than not fails with a `GdsApi::TimedOutException`.

To make the `resync:all` task more reliable, the Resync service that it
calls has been wrapped in a Sidekiq job so that it can be retried.