Add first published at check
This adds a new concept of time_problems to the integrity checker, the
first time to be checked is the first_published_at time. This is the
time that the document is first published regardless of the edition.
whitehall and content-publisher send nil for this value before the
document is first published.

We use first_public_at as this is deemed more reliable than
first_published_at even though they should be the same.

first_public_at time is returned from publishing-api in the format -
"2020-03-11T12:00:00.000+00:00" and so the specs are using this same
format for the pub-api edition.