Add check for public_updated_at
We now want to add a check to make sure public_updated_at values match
between content-publisher and publishing-api.

There is some weirdness here regarding the differences between
content-publishers public_updated_at time and whitehalls regarding first
drafts - whitehall uses the time the edition was updated locally and
then when the edition is published where as content-publisher doesn't
set the time explicitly and generates it on the fly whenever the payload is
generated to be sent to publishing-api. These times will always be
different however knowing the time doesn't actually matter at this stage
because it will be rewritten at the time of publishing I've decided to
skip this check for the first draft edition.

The publishing-api returns public_updated_at in the format -
"2020-03-11T12:00:00Z", this is reflected in the specs.