Remove number of pages from payload if nil
The schema for `publication_attachment_asset` which is used to populated
`details.attachments` in the content item, expects the value for
`number_of_pages` to be an integer:

https://github.com/alphagov/govuk-content-schemas/blob/ac8fb929789c3678e420f7e58db7eb2519692ac0/dist/formats/publication/notification/schema.json#L858

However, when an image is added as an inline attachment, the value is
being set to `nil`. This means that when we try to update the document
in publishing-api it's throwing an error:

https://sentry.io/organizations/govuk/issues/999142960/?project=1242052&referrer=slack

This change excludes `number_of_pages` from the payload if it doesn't
have a value.