Perform integrity check on the change_history during migration
- Update the integrity checker to ensure that the change notes we are replacing match the ones in the Publishing API

We are excluding the match of the first public_timestamp for draft documents.

This is because both Whitehall and Content Publisher allow users to backdate content, to make it appear as if it were published at an earlier time.
The backdated_at field on the edition stores this timestamp, and it is used to populate the public_timestamp value in the change history for the first (draft) edition.

This feature is only restricted to the first edition of the document, due to Publishing API’s approach to change notes. If content publisher allowed backdating after the first edition, then there could be past change notes are newer than the first publishing date, which would not be chronological.

Therefore, if the public_timestamp for a first draft value differs slightly, then it is not a major concern to us.