Format times for PublishingApiPayload to rfc3339
Converts time objects to rfc3339 strings for the generated payload's
timestamps. The reason we've decided to do this is that when we come to
compare times it helps if we're dealing with the same objects, e.g all
strings. An example of where we are comparing times is the
IntegrityChecker where we fetch the edition from the publishing-api and the
timestamps that are returned are strings where as the payload generates
Time objects.

Doing this will make comparing these two times in the IntegrityChecker
simpler for future pieces of work.